remember lucene index dir path, then on delete delete all index parent dirs

+++ core/src/main/java/com/orientechnologies/common/io/OFileUtils.java
@@ -148,7 +148,11 @@ public class OFileUtils {
       throw new IOException("Invalid file name '" + iFileName + "'");
   }
 
-  public static void deleteRecursively(final File rootFile) {
+  public static void deleteRecursively(final File rootFile){
+    deleteRecursively(rootFile, false);
+  }
+  
+  public static void deleteRecursively(final File rootFile, boolean onlyDirs) {
     if (!rootFile.exists())
       return;
 
@@ -157,8 +161,10 @@ public class OFileUtils {
       Files.walkFileTree(rootPath, new SimpleFileVisitor<Path>() {
         @Override
         public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-          if (file != null && file.toFile() != null && file.toFile().exists()){
-            file.toFile().delete();
+          if (!onlyDirs){
+            if (file != null && file.toFile() != null && file.toFile().exists()){
+              file.toFile().delete();
+            }
           }
           return FileVisitResult.CONTINUE;
         }
